Changes to segragate long running performance test into separate profile
